Minor refactor, removed repeating code
